add get a list of all orders
[#160552987]
